Add cisco specific config for qemu.conf
Right now, the cisco packages use an older
version of the libvirt module than cloud-archive.

This older package required additional configuration
to the qemu.conf file.

Currently, this configuration is being done by setting
virtual resources that are declared in the quantum class.

It is worth noting that those resources should not even
be in the quantum config, so this code will need to
be removed ASAP.