inherit params from build_server roles
Previously, the build_server role was using params
as class param defaults, but not inheriting them (and
thus not initializing the params class namespace)

This resulted in the enable_cobbler and enable_cache
parameters always being evaluated as false.

This commit adds the inheritance of params so that things
are evaluated correctly.