Add apache profile
This commit adds an apache profile which serves
as a central location to configure apache so that
it can be configured, and still used as a dependency
for multiple services.

This change is required b/c of 41b0a42b07f6681e01363d7d5738e5f46d747f64
from naginator which includes apache::mod::php which requires
that apache is configured to use prefork to manage processes.