Remove coi::profiles::apache include and migrate to common.yaml
The apache.pp profile included a parameter which made combining
the build profile with all_in_one difficult.  This change removes
the parameter definition and adds it to hiera_data/common.yaml.

The hiera_data/common.yaml must now include:

apache::mpm_module: prefork

which was the parameter being set. A separate update for
openstack-installer will be posted to add that parameter.