Move rubygem install into cobbler_server.pp
Previously we installed rubygems via a static line in a static
preseed file.  This commit consolidates package installs by
putting rubygems in the same place in the manifests as other
packages.  The entry in the static file (and indeed the static
file itself) will be removed in a subsequent commit to
puppet-cobbler.  Until that commit goes in, this commit hurts
nothing and is merely a bit redundant.

Partial-Bug: #1263310