Don't install puppet until Cisco repo is set up
Due to a change in puppet-cobbler to fix an issue with repository
definition duplication, we need to defer the installation
of Puppet until late_command (so that the Cisco repo can be set up
before Puppet is installed).  This patch does that by removing
Puppet from the install list and letting puppet-cobbler handle
it in late_command.

Partial-Bug: #1265850