Restore ability to set custom kernel and cmdopts
In the Grizzly we provided an option in site.pp to allow users
to specify that a custom kernel should be loaded on their nodes
and booted by default.  We also provided an option to specify
custom command line options to be fed to the kernel.  When we moved
to the new data model, those options were (temporarily) lost.  This
commit restores them by providing a hiera lookup and corresponding
actions affeting the preseed file.

Partial-Bug: #1263310