Fix erroneous hiera lookup affecting pip
TL;DR this patch fixes an erroneous hiera lookup caused by a
variable name change between Grizzly and Havana.  We're doing a
lookup for 'default_gateway' when in fact the yaml data uses
'node_gateway'.  This impacts the setup for pip as it will always
see the variable as undef.  More info on the history of this issue
at: https://bugs.launchpad.net/openstack-cisco/+bug/1267942

Closes-Bug: #1267942