Add ability to autoload kernel modules
This commit adds a parameter allowing the user to specify a list of
kernel modules to be added /etc/modules (and thus loaded into the
kernel automatically at boot time).  The default list of modules
includes the 8021q module which was added by default in previous
renditions of this module and the vhost_net module, which has been
requested to improve networking performance.

Implements: blueprint custom-modules-in-baremetal-provisioning