Use hiera lookup to set Cisco package repo info
Previous versions of the puppet-coi module didn't have any
parameters related to setting up the Cisco apt repos in the
Cobbler preseed file's late_command.  Instead, they relied on
a static string in the preseed template.  This meant that the
template, along with code in other modules, needed to be touched
with each release in order to set it to a string that made sense
for that particular release.  This violates DRY and leads to
either release operations or end-user errors.  This patch adds
params for four parameters that are used to supply all
the necessary information to setup the repositories in the preseed,
and passes them on to the Cobbler module.  We also check for and
correct a condition present in h.0 in which the URL strings
specified in data/hiera_data/vendor/cisco_coi_common.yaml and/or
data/hiera_data/enable_ha/true.yaml aren't of the same format,
which could cause h.0 clients to break if they acquire this change
due to Bug 1269856.

Partial-Bug: #1269856