Parameterize puppetdb port and change default
PuppetDB listens on port 8080 out of the box, which conflicts
with the Swift proxy service.  This patch parameterizes the
setting (and also the SSL port) and sets it to a nonconflicting
port (8083) by default.  PuppetDB is also set to listen on
an FQDN which, for the AIO scenario, resolves to a link local
address.  This patch also creates a parameter for the IP
on which PuppetDB should listen and defaults it to 0.0.0.0.

Closes-Bug: #1277728